Add InMemory ShopOrder persistence

+++ usecase/src/main/order/access/shop_order_extractor.rs
@@ -7,5 +7,5 @@ use domain::main::order::value_objects::shop_order_id::ShopOrderId;
 pub trait ShopOrderExtractor: Debug + Send {
     fn get_by_id(&mut self, order_id: ShopOrderId) -> Option<ShopOrder>;
     fn get_last_order(&mut self, for_customer: CustomerId) -> Option<ShopOrder>;
-    fn get_all(&mut self, start_id: ShopOrderId, limit: i32) -> Vec<ShopOrder>;
+    fn get_all(&mut self, start_id: ShopOrderId, limit: usize) -> Vec<ShopOrder>;
 }
